backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.17  # FastAPI Form/File (D5 통합)
httpx==0.27.2

# AI providers
anthropic==0.39.0
google-generativeai==0.8.3   # legacy SDK: embedding (gemini-embedding-001)
google-genai>=1.73           # new SDK: TTS (gemini-2.5-flash-preview-tts), audio response
# openai 미사용 (ADR-008: Gemini 채택)

# Data / DB
# supabase >=2.15: 새 API key 포맷 (sb_publishable_/sb_secret_, 2025-10~) 호환
supabase==2.29.0
pgvector==0.3.6
pandas==2.2.3
rapidfuzz==3.10.1  # price_sentinel fuzzy 매칭 (D4)

# Image processing
Pillow==11.0.0

# TTS
# - Edge TTS: free, no key, unlimited quota (primary)
# - Gemini TTS: paid quota fallback for premium prosody
edge-tts==7.2.8
# google-cloud-texttospeech==2.21.0  # not used (ADR-009)

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0

# Dev tools
black==24.10.0
ruff==0.7.1
